helper/requirements.txt
oss2
requests
click
loguru
pyside6
matplotlib
pyppeteer
python-dotenv